fully organized the esri map
