+++ lib/ui/esrimap/esrimap.dart
@@ -7,19 +7,22 @@
 
 import 'dart:async';
 import 'package:arcgis_maps/arcgis_maps.dart';
-import 'package:campus_mobile_experimental/ui/esrimap/esrimap_ai_search.dart';
-import 'package:campus_mobile_experimental/ui/esrimap/esrimap_basemaps.dart';
-import 'package:campus_mobile_experimental/ui/esrimap/esrimap_config.dart';
-import 'package:campus_mobile_experimental/ui/esrimap/esrimap_fab.dart';
-import 'package:campus_mobile_experimental/ui/esrimap/esrimap_layers_panel.dart';
-import 'package:campus_mobile_experimental/ui/esrimap/esrimap_scene.dart';
+import 'package:campus_mobile_experimental/ui/esrimap/models/esrimap_ai_search_model.dart';
+import 'package:campus_mobile_experimental/ui/esrimap/models/esrimap_basemaps.dart';
+import 'package:campus_mobile_experimental/ui/esrimap/models/esrimap_config.dart';
 import 'package:campus_mobile_experimental/ui/esrimap/models/esrimap_search_category.dart';
 import 'package:campus_mobile_experimental/ui/esrimap/models/map_search_result.dart';
 import 'package:campus_mobile_experimental/ui/esrimap/services/esrimap_auth_handler.dart';
+import 'package:campus_mobile_experimental/ui/esrimap/services/esrimap_basemap_factory.dart';
+import 'package:campus_mobile_experimental/ui/esrimap/services/esrimap_config_service.dart';
 import 'package:campus_mobile_experimental/ui/esrimap/services/esrimap_route_service.dart';
 import 'package:campus_mobile_experimental/ui/esrimap/services/esrimap_search_service.dart';
+import 'package:campus_mobile_experimental/ui/esrimap/widgets/esrimap_ai_search_sheet.dart';
 import 'package:campus_mobile_experimental/ui/esrimap/widgets/esrimap_category_list_panel.dart';
 import 'package:campus_mobile_experimental/ui/esrimap/widgets/esrimap_detail_slide_over.dart';
+import 'package:campus_mobile_experimental/ui/esrimap/widgets/esrimap_fab.dart';
+import 'package:campus_mobile_experimental/ui/esrimap/widgets/esrimap_layers_panel.dart';
+import 'package:campus_mobile_experimental/ui/esrimap/widgets/esrimap_scene.dart';
 import 'package:campus_mobile_experimental/ui/esrimap/widgets/esrimap_search_bar.dart';
 import 'package:campus_mobile_experimental/ui/esrimap/widgets/esrimap_suggestions_panel.dart';
 import 'package:flutter/material.dart';
